smart_budget_buddy/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
streamlit
google-generativeai
plotly
